Add modular onboarding validation components
